requirements/production.txt
arrow==1.3.0
asgiref==3.7.2
blessed==1.20.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.4
Django==4.2.10
django-cors-headers==4.3.1
django-filter==23.5
django-jazzmin==2.6.0
django-picklefield==3.1
django-q==1.3.9
django-safedelete==1.3.3
django-unfold==0.20.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
Markdown==3.5.2
mysqlclient==2.2.4
ndg-httpsclient==0.5.1
psycopg2-binary==2.9.9
pusher==3.3.2
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
pymongo==3.13.0
PyNaCl==1.5.0
pyOpenSSL==24.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
redis==3.5.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
types-python-dateutil==2.8.19.20240106
urllib3==2.2.1
wcwidth==0.2.13
